feat: separate fetch worker from smartWallet worker
 - use ERef<RpcClient>, EREf(LCD)
